Refactor RemoteCacheManager into wildfly-clustering-infinispan-client
