+++ clustering/infinispan/client/src/main/java/org/wildfly/clustering/infinispan/client/manager/RemoteCacheManager.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.clustering.infinispan.subsystem.remote;
+package org.wildfly.clustering.infinispan.client.manager;
 
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
